Fixing unsigned 64-Bit Int 'chat_id' value parsing from .json configuration file; Extending support for Optional parameters.

+++ Kufar Telegram Notifier/telegram.hpp
@@ -11,7 +11,7 @@
 namespace Telegram {
     struct TelegramConfiguration {
         std::string botToken;
-        unsigned int chatID;
+        uint64_t chatID;
     };
 
     void sendAdvert(const TelegramConfiguration &, const Kufar::Ad &);
